+++ core/src/main/java/com/nabla/sdk/core/data/apollo/CoreGqlMapper.kt
@@ -3,10 +3,10 @@ package com.nabla.sdk.core.data.apollo
 import com.nabla.sdk.core.annotation.NablaInternal
 import com.nabla.sdk.core.domain.boundary.Logger
 import com.nabla.sdk.core.domain.entity.EphemeralUrl
-import com.nabla.sdk.core.domain.entity.LivekitRoom
-import com.nabla.sdk.core.domain.entity.LivekitRoomStatus
 import com.nabla.sdk.core.domain.entity.Provider
 import com.nabla.sdk.core.domain.entity.Uri
+import com.nabla.sdk.core.domain.entity.VideoCallRoom
+import com.nabla.sdk.core.domain.entity.VideoCallRoomStatus
 import com.nabla.sdk.graphql.fragment.EphemeralUrlFragment
 import com.nabla.sdk.graphql.fragment.LivekitRoomFragment
 import com.nabla.sdk.graphql.fragment.LivekitRoomStatusFragment
@@ -15,25 +15,25 @@ import com.nabla.sdk.graphql.fragment.ProviderFragment
 @NablaInternal
 public class CoreGqlMapper(private val logger: Logger) {
 
-    public fun mapToLivekitRoom(fragment: LivekitRoomFragment): LivekitRoom {
-        return LivekitRoom(
+    public fun mapToVideoCallRoom(fragment: LivekitRoomFragment): VideoCallRoom {
+        return VideoCallRoom(
             id = fragment.uuid,
-            status = mapToLivekitRoomStatus(fragment.status.livekitRoomStatusFragment),
+            status = mapToVideoCallRoomStatus(fragment.status.livekitRoomStatusFragment),
         )
     }
 
-    public fun mapToLivekitRoomStatus(fragment: LivekitRoomStatusFragment): LivekitRoomStatus {
+    public fun mapToVideoCallRoomStatus(fragment: LivekitRoomStatusFragment): VideoCallRoomStatus {
         fragment.onLivekitRoomOpenStatus?.let {
-            return LivekitRoomStatus.Open(
+            return VideoCallRoomStatus.Open(
                 url = it.url,
                 token = it.token
             )
         }
         fragment.onLivekitRoomClosedStatus?.let {
-            return LivekitRoomStatus.Closed
+            return VideoCallRoomStatus.Closed
         }
-        logger.error("Unknown livekit room status mapping for $fragment - falling back to Closed")
-        return LivekitRoomStatus.Closed
+        logger.error("Unknown video call room status mapping for $fragment - falling back to Closed")
+        return VideoCallRoomStatus.Closed
     }
 
     public fun mapToProvider(providerFragment: ProviderFragment): Provider {