+++ core/src/main/java/com/nabla/sdk/core/domain/entity/VideoCallRoom.kt
@@ -2,7 +2,7 @@ package com.nabla.sdk.core.domain.entity
 
 import com.benasher44.uuid.Uuid
 
-public data class LivekitRoom(
+public data class VideoCallRoom(
     val id: Uuid,
-    val status: LivekitRoomStatus
+    val status: VideoCallRoomStatus
 )