+++ messaging-core-test-fixtures/src/main/java/com/nabla/sdk/messaging/core/data/stubs/ConversationContentRepositoryStub.kt
@@ -227,7 +227,7 @@ internal class ConversationContentRepositoryStub(
             is Message.Media.Image -> copy(baseMessage.copy(sendStatus = newStatus, id = newId))
             is Message.Media.Video -> copy(baseMessage.copy(sendStatus = newStatus, id = newId))
             is Message.Text -> copy(baseMessage.copy(sendStatus = newStatus, id = newId))
-            is Message.LivekitRoom -> copy(baseMessage.copy(sendStatus = newStatus, id = newId))
+            is Message.VideoCallRoom -> copy(baseMessage.copy(sendStatus = newStatus, id = newId))
         }
     }
 }