+++ scheduling/src/main/java/com/nabla/sdk/scheduling/scene/appointments/ItemUiModel.kt
@@ -1,9 +1,9 @@
 package com.nabla.sdk.scheduling.scene.appointments
 
 import com.benasher44.uuid.Uuid
-import com.nabla.sdk.core.domain.entity.LivekitRoom
-import com.nabla.sdk.core.domain.entity.LivekitRoomStatus
 import com.nabla.sdk.core.domain.entity.Provider
+import com.nabla.sdk.core.domain.entity.VideoCallRoom
+import com.nabla.sdk.core.domain.entity.VideoCallRoomStatus
 import com.nabla.sdk.scheduling.domain.entity.Appointment
 import com.nabla.sdk.scheduling.domain.entity.AppointmentId
 import com.nabla.sdk.scheduling.scene.appointments.ItemUiModel.AppointmentUiModel.SoonOrOngoing.CallButtonStatus
@@ -33,10 +33,10 @@ internal sealed class ItemUiModel(val listId: String) {
             sealed interface CallButtonStatus {
                 object Absent : CallButtonStatus
                 sealed interface Present : CallButtonStatus {
-                    val livekitRoom: LivekitRoom
+                    val videoCallRoom: VideoCallRoom
 
-                    data class AsJoin(override val livekitRoom: LivekitRoom) : Present
-                    data class AsGoBack(override val livekitRoom: LivekitRoom) : Present
+                    data class AsJoin(override val videoCallRoom: VideoCallRoom) : Present
+                    data class AsGoBack(override val videoCallRoom: VideoCallRoom) : Present
                 }
             }
         }
@@ -62,18 +62,18 @@ internal fun Appointment.toUiModel(currentCallId: Uuid?) = when (this) {
             id,
             provider,
             scheduledAt,
-            callButtonStatus = when (livekitRoom?.status) {
-                is LivekitRoomStatus.Open -> {
+            callButtonStatus = when (videoCallRoom?.status) {
+                is VideoCallRoomStatus.Open -> {
                     when (currentCallId) {
-                        null -> CallButtonStatus.Present.AsJoin(livekitRoom)
-                        livekitRoom.id -> CallButtonStatus.Present.AsGoBack(livekitRoom)
+                        null -> CallButtonStatus.Present.AsJoin(videoCallRoom)
+                        videoCallRoom.id -> CallButtonStatus.Present.AsGoBack(videoCallRoom)
                         else -> {
                             // there's a current call but it's not this one — don't show a join until the current call has ended.
                             CallButtonStatus.Absent
                         }
                     }
                 }
-                null, LivekitRoomStatus.Closed -> {
+                null, VideoCallRoomStatus.Closed -> {
                     CallButtonStatus.Absent
                 }
             }